+++ src/Editor/Windows/CodeWindow.cs
@@ -1,19 +1,37 @@
 ﻿using System.Drawing;
 using ImGuiNET;
-using System.Numerics;
 using Editor.Configuration;
 using Library;
 using Library.Packaging;
+using ImGuiColorTextEditNet.Syntax;
+using ImGuiColorTextEditNet;
 
 namespace Editor.Windows
 {
     /// <summary>
     /// Handles the display and the modification of code
     /// </summary>
-    internal class CodeWindow(
-        EditorConfiguration editorConfiguration,
-        EditorControllerData editorControllerData)
+    internal class CodeWindow
     {
+        private readonly EditorControllerData _editorControllerData;
+
+        private Dictionary<FileId, TextEditor> _textEditors = [];
+
+        /// <summary>
+        /// Handles the display and the modification of code
+        /// </summary>
+        public CodeWindow(EditorConfiguration editorConfiguration,
+            EditorControllerData editorControllerData)
+        {
+            _editorControllerData = editorControllerData;
+
+            //_textEditor = new TextEditor
+            //{
+            //    //AllText = demoText,
+            //    SyntaxHighlighter = new CStyleHighlighter(true),
+            //};
+        }
+
         public event Action? BuildPressed;
 
         /// <summary>
@@ -25,7 +43,27 @@ namespace Editor.Windows
         {
             var codeChanged = false;
 
-            editorControllerData.UpdateWindowPosAndSize(EditorControllerData.WindowId.Code);
+            foreach (var (key, code) in shaderCodes)
+            {
+                if (_textEditors.ContainsKey(key) == false)
+                {
+                    var editor = new TextEditor()
+                    {
+                        AllText = code.Code,
+                        SyntaxHighlighter = new GlSlStyleHighlighter()
+                    };
+                    _textEditors.Add(key, editor);
+                }
+            }
+
+            foreach (var key in _textEditors.Keys)
+            {
+                if (shaderCodes.ContainsKey(key) == false)
+                    _textEditors.Remove(key);
+            }
+
+
+            _editorControllerData.UpdateWindowPosAndSize(EditorControllerData.WindowId.Code);
 
             if (ImGui.Begin("Code"))
             {
@@ -46,8 +84,7 @@ namespace Editor.Windows
                 ImGui.SameLine();
                 if (isValid == false)
                     ImGui.TextColored(TypeConverters.ColorToVector4(Color.Red), "not valid, check messages");
-                else
-                if (needsRebuild)
+                else if (needsRebuild)
                     ImGui.TextColored(TypeConverters.ColorToVector4(Color.Orange), "needs rebuild");
                 else
                     ImGui.TextColored(TypeConverters.ColorToVector4(Color.LimeGreen), "valid");
@@ -91,8 +128,7 @@ namespace Editor.Windows
                 ImGui.PushStyleColor(ImGuiCol.TabHovered, TypeConverters.ColorToVector4(Color.IndianRed));
                 styleAttributesPushed++;
             }
-            else
-            if (code.NeedsRebuild)
+            else if (code.NeedsRebuild)
             {
                 ImGui.PushStyleColor(ImGuiCol.Tab, TypeConverters.ColorToVector4(Color.DarkOrange));
                 styleAttributesPushed++;
@@ -105,16 +141,59 @@ namespace Editor.Windows
 
             if (ImGui.BeginTabItem(name))
             {
-                var inputFlags = ImGuiInputTextFlags.AllowTabInput;
-                if (ImGui.InputTextMultiline("##source",
-                        ref code.Code,
-                        20000,
-                        new Vector2(-1, -1),
-                        inputFlags))
+                var textEditor = _textEditors[fileId];
+                ImGui.Text(
+                    $"Cur:{textEditor.CursorPosition} SEL: {textEditor.Selection.Start} - {textEditor.Selection.End}"
+                );
+
+                // We force line endings to \n to be able to properly detect modifications on text
+                var currentTextWithLineFeeds = textEditor.AllText.ReplaceLineEndings("\n");
+                textEditor.AllText = currentTextWithLineFeeds;
+
+                //var demoErrors = new Dictionary<int, object>
+                //{
+                //    { 1, "Syntax error etc 1" },
+                //    { 10, "Syntax error etc 10" }
+                //};
+                //textEditor.ErrorMarkers.SetErrorMarkers(demoErrors);
+
+                textEditor.Render("EditWindow");
+
+                var newTextWithLineFeeds = textEditor.AllText.ReplaceLineEndings("\n");
+
+                if (newTextWithLineFeeds != currentTextWithLineFeeds)
                 {
+                    //var src = newTextWithLineFeeds.ToCharArray();
+                    //var dst = currentTextWithLineFeeds.ToCharArray();
+
+                    //for (int i = 0; i < src.Length; i++)
+                    //{
+                    //    var s = src[i];
+                    //    var d = dst[i];
+
+                    //    if (s != d)
+                    //    {
+                    //        code.NeedsRebuild = true;
+                    //    }
+                    //}
+
                     code.NeedsRebuild = true;
                     codeChanged = true;
                 }
+
+                code.Code = textEditor.AllText;
+
+
+                //var inputFlags = ImGuiInputTextFlags.AllowTabInput;
+                //if (ImGui.InputTextMultiline("##source",
+                //        ref code.Code,
+                //        20000,
+                //        new Vector2(-1, -1),
+                //        inputFlags))
+                //{
+                //    code.NeedsRebuild = true;
+                //    codeChanged = true;
+                //}
 
                 ImGui.EndTabItem();
             }