+++ CHANGELOG.md
@@ -7,4 +7,5 @@
 ### Known bugs
 
 * Switching from 4 to 1 light doesn't clear extra lights in shader
-
+* When using ImGuiColorTextEditNet, SimpleTrie.cs is needed to build the Editor 
+* backspace inside ImGuiColorTextEditNet control raises an exception